realtime-diabetes-prediction/requirements.txt
# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ─── Cache và Message Queue ───────────────────────────────────────────────────
redis==5.0.1
celery==5.3.6

# ─── Xác thực ─────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── Machine Learning ─────────────────────────────────────────────────────────
scikit-learn==1.4.0
tensorflow==2.15.0
shap==0.44.0
imbalanced-learn==0.11.0

# ─── Data Processing ──────────────────────────────────────────────────────────
pandas==2.1.4
numpy==1.26.3

# ─── API Schema ───────────────────────────────────────────────────────────────
pydantic==2.5.3
pydantic[email]==2.5.3
python-dotenv==1.0.0

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.26.0

# ─── Firebase ─────────────────────────────────────────────────────────────────
firebase-admin==6.3.0
